+++ packages/core/src/components/discord-verified-author-tag/DiscordVerifiedAuthorTag.ts
@@ -35,6 +35,8 @@ export class DiscordVerifiedAuthorTag extends LitElement {
 			padding-left: 10px;
 			padding-right: 4px;
 			margin-right: 0.25rem;
+			margin-left: 0px !important;
+			margin-top: 0px !important;
 		}
 
 		:host([compact-mode]) .discord-application-tag-verified {